desktop/requirements.txt
customtkinter==5.2.1
mysql-connector-python==8.2.0
bcrypt==4.1.2
pycryptodome==3.19.0
python-dotenv==1.0.0
Pillow==10.1.0
pyperclip==1.8.2